FEAT(mvp/+): Condense word array into object and store in book model
It seems dumb to store the word list as a raw array when what we'll most often want is to know how many unique words there are, and how many times each is used. So words are now stored as an object where each word is a key whose value is the number of times it was used.

Also now updating the wordCount field to match the length of the token array (the numbers in the seed file are BBEdit's count, which might differ), and added a uniqueCount field to store the number of keys in the map.

Both the GET and POST `/api/books/:bookId/words` and `/api/books/:bookId/sentences` routes return a JSON word/sentence list, and post the list to the database if it's not already there. The only difference betwen GET and POST is what status message you get.