FEAT(mvp/+): Revise word mapper to omit numbers
Also removed the metadata from the top of the home-grown corpus texts. Removing any bracketed stuff at the top of the file might be a better solution, since it doesn't make sense to manually edit the node-gutenberg texts, as they'll just get replaced on the next `npm install`. Shouldn't be hard to write a RegExp to do this—e.g.,
`/^\[.*?\]/`
should work.